# FE/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ChangeMind API",
    "description": "ASP.NET Core 10.0 API following Clean Architecture principles. Provides endpoints for fitness coaching, training programs, and personal health management.",
    "version": "1.0.0",
    "contact": {
      "name": "ChangeMind Team",
      "email": "ongguzel@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development (Gateway)"
    },
    {
      "url": "https://api.changemind.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/nutrition-plans": {
      "post": {
        "tags": ["Nutrition Plans"],
        "summary": "Create a nutrition plan (Coach or Admin)",
        "description": "Creates a nutrition plan covering one or two day types (WorkoutDay and/or OffDay). Any previously active plan for the user is automatically deactivated (single-active rule).",
        "operationId": "CreateNutritionPlan",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateNutritionPlanRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan created",
            "content": { "application/json": { "schema": { "type": "string", "format": "uuid" } } }
          },
          "400": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationProblemDetails" } } }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } },
          "403": { "description": "Forbidden — coach or admin role required", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/api/nutrition-plans/{planId}": {
      "get": {
        "tags": ["Nutrition Plans"],
        "summary": "Get nutrition plan by ID",
        "description": "Users can only view their own plan; coaches and admins can view any plan.",
        "operationId": "GetNutritionPlanById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "planId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Plan retrieved",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NutritionPlanDetailDto" } } }
          },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } },
          "404": { "description": "Plan not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      },
      "put": {
        "tags": ["Nutrition Plans"],
        "summary": "Update nutrition plan (Coach or Admin)",
        "operationId": "UpdateNutritionPlan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "planId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateNutritionPlanRequest" }
            }
          }
        },
        "responses": {
          "204": { "description": "Updated" },
          "404": { "description": "Plan not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/api/nutrition-plans/{planId}/activate": {
      "post": {
        "tags": ["Nutrition Plans"],
        "summary": "Activate nutrition plan (Coach or Admin)",
        "description": "Activates the plan; any other active plans for the same user are deactivated.",
        "operationId": "ActivateNutritionPlan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "planId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Activated" },
          "404": { "description": "Plan not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/api/nutrition-plans/{planId}/deactivate": {
      "post": {
        "tags": ["Nutrition Plans"],
        "summary": "Deactivate nutrition plan (Coach or Admin)",
        "operationId": "DeactivateNutritionPlan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "planId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Deactivated" },
          "404": { "description": "Plan not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/api/users/{userId}/active-nutrition-plan": {
      "get": {
        "tags": ["Nutrition Plans"],
        "summary": "Get the user's active nutrition plan (Coach > Self priority)",
        "description": "Returns a single active nutrition plan for the user. If a Coach-assigned active plan exists, it is returned; otherwise the Self-created active plan is returned (if any). Returns 404 when the user has neither.",
        "operationId": "GetUserActiveNutritionPlan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Plan retrieved",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NutritionPlanDetailDto" } }
            }
          },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } },
          "404": { "description": "No active plan", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
...
    "/api/nutrition-plans/{planId}/export": {
      "get": {
        "tags": ["Nutrition Plans"],
        "summary": "Export a nutrition plan as Excel (.xlsx)",
        "description": "Returns the plan as an .xlsx file. Each NutritionDayType is rendered on its own sheet; meals are stacked vertically. Users may only export their own plans; coaches/admins may export any.",
        "operationId": "ExportNutritionPlan",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "planId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Excel file stream",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } },
          "404": { "description": "Plan not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
    "/api/users/{userId}/nutrition-plans": {
      "get": {
        "tags": ["Nutrition Plans"],
        "summary": "List all nutrition plans (active + historical) for a user (Coach or Admin)",
        "operationId": "GetUserNutritionPlans",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Plans retrieved",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/NutritionPlanListItemDto" } }
              }
            }
          },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://tools.ietf.org/html/rfc7807"
          },
          "title": {
            "type": "string",
            "example": "Not Found"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "User with ID 'xxx' not found."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/users/550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "example": "Validation Failed"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "One or more validation errors occurred."
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "errors": {
            "type": "object",
            "description": "Validation errors indexed by field name or index",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": ["Email is required", "Email format is invalid"],
              "password": ["Password must be at least 8 characters"]
            }
          }
        }
      }
...
    "NutritionDayType": {
      "type": "string",
      "enum": ["WorkoutDay", "OffDay"]
    },
    "MealItemInput": {
      "type": "object",
      "required": ["foodId"],
      "description": "Exactly one of `grams` or `pieces` must be positive, matching the referenced food's `unit`. Grams-based foods accept `grams` only; Piece-based foods accept `pieces` only.",
      "properties": {
        "foodId": { "type": "string", "format": "uuid" },
        "grams":  { "type": "number", "format": "decimal", "minimum": 0, "nullable": true, "example": 150, "description": "Quantity in grams. Required for Grams-based foods." },
        "pieces": { "type": "number", "format": "decimal", "minimum": 0, "nullable": true, "example": 2,   "description": "Quantity in pieces (adet). Required for Piece-based foods." }
      }
    },
    "MealInput": {
      "type": "object",
      "required": ["name", "items"],
      "properties": {
        "name": { "type": "string", "example": "Breakfast" },
        "items": { "type": "array", "items": { "$ref": "#/components/schemas/MealItemInput" } }
      }
    },
    "CreateNutritionPlanRequest": {
      "type": "object",
      "required": ["userId", "title", "days"],
      "properties": {
        "userId": { "type": "string", "format": "uuid" },
        "title": { "type": "string", "maxLength": 200, "example": "Cutting Plan - Week 1" },
        "description": { "type": "string", "maxLength": 1000, "nullable": true },
        "days": {
          "type": "object",
          "description": "Dictionary keyed by NutritionDayType (WorkoutDay / OffDay). One or two entries required.",
          "additionalProperties": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MealInput" }
          },
          "example": {
            "WorkoutDay": [
              { "name": "Breakfast", "items": [ { "foodId": "00000000-0000-0000-0000-0000000F0005", "grams": 100 } ] }
            ],
            "OffDay": [
              { "name": "Lunch", "items": [ { "foodId": "00000000-0000-0000-0000-0000000F0001", "grams": 200 } ] }
            ]
          }
        }
      }
    },
...
    "UpdateNutritionPlanRequest": {
      "type": "object",
      "required": ["title", "days"],
      "properties": {
        "title": { "type": "string", "maxLength": 200 },
        "description": { "type": "string", "maxLength": 1000, "nullable": true },
        "days": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MealInput" }
          }
        }
      }
    },
    "MealItemDto": {
      "type": "object",
      "properties": {
        "foodId":       { "type": "string", "format": "uuid" },
        "foodName":     { "type": "string" },
        "quantity":     { "type": "number", "format": "decimal", "description": "Amount expressed in `quantityUnit`." },
        "quantityUnit": { "type": "string", "enum": ["g", "adet"], "description": "Unit of the `quantity` field; 'g' for grams or 'adet' for pieces." },
        "calories":     { "type": "number", "format": "decimal" },
        "protein":      { "type": "number", "format": "decimal" },
        "carbs":        { "type": "number", "format": "decimal" },
        "fat":          { "type": "number", "format": "decimal" }
      }
    },
    "MealDto": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "items": { "type": "array", "items": { "$ref": "#/components/schemas/MealItemDto" } },
        "totalCalories": { "type": "number", "format": "decimal" },
        "totalProtein": { "type": "number", "format": "decimal" },
        "totalCarbs": { "type": "number", "format": "decimal" },
        "totalFat": { "type": "number", "format": "decimal" }
      }
    },
    "NutritionDayDto": {
      "type": "object",
      "properties": {
        "dayType": { "$ref": "#/components/schemas/NutritionDayType" },
        "meals": { "type": "array", "items": { "$ref": "#/components/schemas/MealDto" } },
        "totalCalories": { "type": "number", "format": "decimal" },
        "totalProtein": { "type": "number", "format": "decimal" },
        "totalCarbs": { "type": "number", "format": "decimal" },
        "totalFat": { "type": "number", "format": "decimal" }
      }
    },
    "NutritionPlanDetailDto": {
      "type": "object",
      "properties": {
        "id": { "type": "string", "format": "uuid" },
        "userId": { "type": "string", "format": "uuid" },
        "coachId": { "type": "string", "format": "uuid", "nullable": true, "description": "Coach ID if `createdByType=Coach`, otherwise null." },
        "coachName": { "type": "string", "nullable": true, "description": "Null if `createdByType=Self`." },
        "createdByType": { "type": "string", "enum": ["Self", "Coach"] },
        "title": { "type": "string" },
        "description": { "type": "string", "nullable": true },
        "isActive": { "type": "boolean" },
        "versionNumber": { "type": "integer", "format": "int32" },
        "createdAt": { "type": "string", "format": "date-time" },
        "updatedAt": { "type": "string", "format": "date-time", "nullable": true },
        "days": { "type": "array", "items": { "$ref": "#/components/schemas/NutritionDayDto" } }
      }
    },
    "NutritionPlanListItemDto": {
      "type": "object",
      "properties": {
        "id": { "type": "string", "format": "uuid" },
        "title": { "type": "string" },
        "isActive": { "type": "boolean" },
        "coachName": { "type": "string" },
        "createdAt": { "type": "string", "format": "date-time" }
      }
    }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtained from /api/auth/login endpoint. Token validity: 15 minutes. Use Idempotency-Key header with POST /api/payments for idempotent payment processing."
      }
    }
  },
...
}